feat: add of modal base
